corrected modes grouping and other fixes
